CA-Q10- Lstq Finished
Completed CA-Q10. That was last CA question. I used mapding and lambda expression to map corresponding elements to each other and subtract them in the process. Outcome of this is a singular list with elements that are already subtracted values. I have used square function to multiply the elements by themselves, while recursively looping through remaining values of that list adding the outcome of squared function each time, which results in a single number output